feat : 랜덤 중복 숫자 제거 기능 구현
