+++ src/main/java/game/baesball/util/RandomUtil.java
@@ -6,11 +6,15 @@ public class RandomUtil {
 
     public int[] makeThreeRandoms() {
         Random random = setRandom();
-        return new int[]{
+        int[] randomNumbers = {
                 getRandomNumberBetweenOnetoNine(random),
                 getRandomNumberBetweenOnetoNine(random),
                 getRandomNumberBetweenOnetoNine(random)
         };
+        if (checkSameNumber(randomNumbers)) {
+            randomNumbers = makeThreeRandoms();
+        }
+        return randomNumbers;
     }
 
     private int getRandomNumberBetweenOnetoNine(Random random) {
@@ -21,5 +25,14 @@ public class RandomUtil {
         Random random = new Random();
         random.setSeed(System.currentTimeMillis());
         return random;
+    }
+
+    private boolean checkSameNumber(int[] numbers) {
+        for (int curr = 0; curr < numbers.length - 1; curr++) {
+            for (int next = curr + 1; next < numbers.length; next++) {
+                if (numbers[curr] == numbers[next]) return true;
+            }
+        }
+        return false;
     }
 }
